Releases/1.0.0 dev

+++ app/src/main/java/com/keith/expensesplitter/data/model/Expense.kt
@@ -17,6 +17,6 @@ data class Expense (
     @PrimaryKey(autoGenerate = true)
     val id: Long? = null,
     val name: String,
-    val amount: Float,
+    val amount: Long,
     val groupId: Long
 )
